implemented RUD

+++ src/main/java/net/proselyte/springsecurityapp/dao/RoleDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by tramon on 26.01.2017.
  */
-public interface RoleDao extends JpaRepository<Role, Long> {
+public interface RoleDao extends JpaRepository<Role, Integer> {
 
 
 }
